Expose IServiceBusMessageFactory and ServiceBusMessageFactory
