+++ source/Messaging/source/Communication/Publisher/IServiceBusMessageFactory.cs
@@ -13,23 +13,13 @@
 // limitations under the License.
 
 using Azure.Messaging.ServiceBus;
-using Google.Protobuf;
 
-namespace Energinet.DataHub.Core.Messaging.Communication.Internal.Publisher;
+namespace Energinet.DataHub.Core.Messaging.Communication.Publisher;
 
-internal sealed class ServiceBusMessageFactory : IServiceBusMessageFactory
+/// <summary>
+/// Creates a <see cref="ServiceBusMessage"/> instance from an <see cref="IntegrationEvent"/>
+/// </summary>
+public interface IServiceBusMessageFactory
 {
-    public ServiceBusMessage Create(IntegrationEvent @event)
-    {
-        var serviceBusMessage = new ServiceBusMessage
-        {
-            Body = new BinaryData(@event.Message.ToByteArray()),
-            Subject = @event.EventName,
-            MessageId = @event.EventIdentification.ToString(),
-        };
-
-        serviceBusMessage.ApplicationProperties.Add("EventMinorVersion", @event.EventMinorVersion);
-
-        return serviceBusMessage;
-    }
+    ServiceBusMessage Create(IntegrationEvent @event);
 }
